requirements.txt
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pandas>=2.2.0
pymupdf>=1.24.0
tqdm>=4.66.0
scikit-learn>=1.5.0
streamlit
Flask
flask-cors
